+++ packages/vchart/src/chart/box-plot/box-plot-transformer.ts
@@ -10,13 +10,19 @@ export class BoxPlotChartSpecTransformer<
     const dataFields = [spec.maxField, spec.medianField, spec.q1Field, spec.q3Field, spec.minField, spec.outliersField];
     const seriesSpec = super._getDefaultSeriesSpec(spec, [
       'boxPlot',
+      'outlier',
       'minField',
       'maxField',
       'q1Field',
       'medianField',
       'q3Field',
       'outliersField',
-      'outliersStyle'
+      'outliersStyle',
+
+      'boxWidth',
+      'boxMaxWidth',
+      'boxMinWidth',
+      'boxGapInGroup'
     ]);
     seriesSpec.direction = spec.direction ?? Direction.vertical;
     seriesSpec[seriesSpec.direction === Direction.horizontal ? 'xField' : 'yField'] = dataFields;